Move react components to their own dedicated directory.
