--- galaxyui/src/app/authors-react/react-components/render.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-
-export class Renderer {
-    static init(injector, component) {
-
-    }
-}
