+++ galaxyui/src/app/react/styles/community.less
