+++ galaxyui/src/app/react/styles/page-header.less
