+++ galaxyui/src/app/react/styles/page-loading.less
