+++ galaxyui/src/styles.less
@@ -260,10 +260,10 @@ a.failed {
 // loaded into JSX
 
 // Page Header
-@import './app/authors-react/react-components/page-header.less';
+@import './app/react/styles/page-header.less';
 
 // Community Page
-@import './app/authors-react/react-components/community.less';
+@import './app/react/styles/community.less';
 
 // Page Loading Widget
-@import './app/authors-react/react-components/page-loading.less';
+@import './app/react/styles/page-loading.less';
